Made the I/O library actually work with the echo server.
We were checking for partial reads/writes and then reposting the
events when that happened. This breaks a lot of network code where
we try to read a MAX_LEN from the socket. Instead make cilk_read and
cilk_write have the same semantics as that of posix read/write where
the application has to handle partial reads/writes.